requirements.txt
python-dotenv
discord.py
schedule
